Lift the parts of a workflow run that decide, not do
Graph shape, retry seeding, loop exit, condition evaluation, fingerprints and
the template resolver: none of it opens a socket, spawns a process or reads a
store, and both sides need it -- the engine walks a run with it, the editor
reads the same answers to draw one.

It moves to `packages/shared` ahead of the engine itself, so that when
execution moves to the server the two cannot answer the same question
differently. Nothing changes behaviourally: the renderer's engine imports what
it used to define, and re-exports what the editor and run views import from it.
