+++ src/renderer/components/SourcePromptDialog.tsx
@@ -4,7 +4,7 @@ import { GitBranch, Zap } from 'lucide-react'
 import { useAppStore } from '../stores'
 import { ProjectPicker } from './ProjectPicker'
 import { executeWorkflow } from '../lib/workflow-execution'
-import { containsContextRef, isContextRef } from '../lib/template-vars'
+import { containsContextRef, isContextRef } from '@vornrun/shared/template-vars'
 import { getWorkflowInputs, isContextualWorkflow } from '../lib/workflow-helpers'
 import { WorkflowInputFields } from './WorkflowInputFields'
 import { areInputsValid, initialInputValues } from '../lib/workflow-inputs'
