+++ src/renderer/components/workflow-editor/panels/LaunchAgentConfigForm.tsx
@@ -14,7 +14,7 @@ import {
   CONTEXT_REF,
   getAvailableContextVars,
   isContextRef
-} from '../../../lib/template-vars'
+} from '@vornrun/shared/template-vars'
 import { getWorktreeMode } from '../../../lib/workflow-helpers'
 import { DEFAULT_STEP_TIMEOUT_MINUTES } from '../../../lib/workflow-execution'
 import { useAgentInstallStatus } from '../../../hooks/useAgentInstallStatus'
