+++ tests/workflow-step-worktree.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
 import { buildStepOutputsMap } from '../src/renderer/lib/workflow-execution'
-import { DEFAULT_OUTPUT_KEYS, WORKTREE_OUTPUT_KEY } from '../src/renderer/lib/template-vars'
+import { DEFAULT_OUTPUT_KEYS, WORKTREE_OUTPUT_KEY } from '@vornrun/shared/template-vars'
 import type { WorkflowExecution, WorkflowNode } from '../src/shared/types'
 
 // The directory a step worked in, offered to the steps after it.
